requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.6.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pgvector==0.2.4
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.1.0
prompt-toolkit==3.0.42
psutil==5.9.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.0
regex==2023.12.25
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
tiktoken==0.5.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
wheel==0.41.2
zipp==3.17.0
